Update: Modularise all Grid-related components
