+++ src/components/adslotUi/PagedGridComponent.jsx
@@ -2,9 +2,9 @@ import _ from 'lodash';
 import React, { PropTypes } from 'react';
 import Pagination from 'react-bootstrap/lib/Pagination';
 import Empty from 'components/alexandria/Empty/component';
-import Grid from 'components/alexandria/GridComponent';
-import GridRow from 'components/alexandria/GridRowComponent';
-import GridCell from 'components/alexandria/GridCellComponent';
+import Grid from 'components/alexandria/Grid/component';
+import GridRow from 'components/alexandria/GridRow/component';
+import GridCell from 'components/alexandria/GridCell/component';
 import SvgSymbol from 'components/alexandria/SvgSymbolComponent';
 
 require('styles/adslotUi/PagedGrid.scss');
