+++ src/components/alexandria/Grid/styles.scss
@@ -1,7 +1,7 @@
-@import '../color';
-@import '../font-weight';
-@import '../font-size';
-@import '../border';
+@import 'styles/color';
+@import 'styles/font-weight';
+@import 'styles/font-size';
+@import 'styles/border';
 
 .grid-component {
   border-left: $border-lighter;
