+++ src/components/alexandria/GridCell/component.jsx
@@ -1,9 +1,7 @@
 import React, { PropTypes } from 'react';
-
-import classSuffixHelper from '../../helpers/classSuffixHelper';
-import expandDts from '../../helpers/expandDtsHelper';
-
-require('styles/alexandria/GridCell.scss');
+import classSuffixHelper from '../../../helpers/classSuffixHelper';
+import expandDts from '../../../helpers/expandDtsHelper';
+import './styles.scss';
 
 const GridCellComponent = ({ children, classSuffixes, onClick, stretch, dts }) => {
   const componentClass = 'grid-component-cell';
