+++ src/components/alexandria/GridCell/styles.scss
@@ -1,4 +1,4 @@
-@import '../mixin';
+@import 'styles/mixin';
 
 $cell-padding: 10px;
