+++ src/components/alexandria/GridRow/styles.scss
@@ -1,6 +1,6 @@
-@import '../color';
-@import '../border';
-@import '../font-weight';
+@import 'styles/color';
+@import 'styles/border';
+@import 'styles/font-weight';
 
 .grid-component-row {
   border-bottom: 1px solid transparent;
