+++ src/components/core.js
@@ -44,9 +44,9 @@ import Breadcrumb from 'components/alexandria/Breadcrumb/component';
 import Card from 'components/alexandria/Card/component';
 import Empty from 'components/alexandria/Empty/component';
 import FlexibleSpacer from 'components/alexandria/FlexibleSpacer/component';
-import Grid from 'components/alexandria/GridComponent';
-import GridCell from 'components/alexandria/GridCellComponent';
-import GridRow from 'components/alexandria/GridRowComponent';
+import Grid from 'components/alexandria/Grid/component';
+import GridCell from 'components/alexandria/GridCell/component';
+import GridRow from 'components/alexandria/GridRow/component';
 import PageTitle from 'components/alexandria/PageTitleComponent';
 import PrettyDiff from 'components/alexandria/PrettyDiffComponent';
 import Search from 'components/alexandria/SearchComponent';
