+++ test/components/adslotUi/ListPickerPureComponentTest.jsx
@@ -5,9 +5,9 @@ import Checkbox from 'react-icheck/lib/Checkbox';
 import Radio from 'react-icheck/lib/Radio';
 import ListPickerPureComponent from 'components/adslotUi/ListPickerPureComponent';
 import Empty from 'components/alexandria/Empty/component';
-import Grid from 'components/alexandria/GridComponent';
-import GridCell from 'components/alexandria/GridCellComponent';
-import GridRow from 'components/alexandria/GridRowComponent';
+import Grid from 'components/alexandria/Grid/component';
+import GridCell from 'components/alexandria/GridCell/component';
+import GridRow from 'components/alexandria/GridRow/component';
 import ListPickerMocks from 'mocks/ListPickerMocks';
 
 describe('ListPickerPureComponent', () => {
