+++ test/components/adslotUi/TreePicker/TreePickerGridComponentTest.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import TreePickerGrid, { TreePickerNodeFast } from 'components/adslotUi/TreePicker/TreePickerGridComponent';
 import Empty from 'components/alexandria/Empty/component';
-import Grid from 'components/alexandria/GridComponent';
+import Grid from 'components/alexandria/Grid/component';
 import TreePickerMocks from 'mocks/TreePickerMocks';
 
 describe('TreePickerGridComponent', () => {
