+++ test/components/adslotUi/fastStatelessWrapperTest.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import fastStatelessWrapper from 'components/adslotUi/fastStatelessWrapper';
-import Grid from 'components/alexandria/GridComponent';
+import Grid from 'components/alexandria/Grid/component';
 
 describe('fastStatelessWrapper', () => {
   it('should re-render on specified attribute change', () => {
